+++ src/adapter/graphql/src/queries/accounts/accounts.rs
@@ -17,7 +17,25 @@ pub struct Accounts;
 #[common_macros::method_names_consts(const_value_prefix = "Gql::")]
 #[Object]
 impl Accounts {
-    /// Returns account by its ID
+    /// Returns the account that is making the call (authorized subject)
+    #[tracing::instrument(level = "info", name = Accounts_me, skip_all)]
+    async fn me(&self, ctx: &Context<'_>) -> Result<Account> {
+        use kamu_accounts::CurrentAccountSubject;
+
+        let subject = from_catalog_n!(ctx, CurrentAccountSubject);
+
+        match subject.as_ref() {
+            CurrentAccountSubject::Logged(account) => Ok(Account::new(
+                account.account_id.clone().into(),
+                account.account_name.clone().into(),
+            )),
+            CurrentAccountSubject::Anonymous(_) => {
+                Err(GqlError::gql(ANONYMOUS_ACCESS_FORBIDDEN_MESSAGE))
+            }
+        }
+    }
+
+    /// Returns an account by its ID, if found
     #[tracing::instrument(level = "info", name = Accounts_by_id, skip_all, fields(%account_id))]
     async fn by_id(&self, ctx: &Context<'_>, account_id: AccountID<'_>) -> Result<Option<Account>> {
         let account_service = from_catalog_n!(ctx, dyn kamu_accounts::AccountService);
@@ -29,7 +47,41 @@ impl Accounts {
             .map(|account_name| Account::new(account_id.into(), account_name.into())))
     }
 
-    /// Returns account by its name
+    /// Returns accounts by their IDs
+    #[tracing::instrument(level = "info", name = Accounts_by_ids, skip_all, fields(?account_ids, ?skip_missing))]
+    async fn by_ids(
+        &self,
+        ctx: &Context<'_>,
+        account_ids: Vec<AccountID<'_>>,
+        #[graphql(
+            desc = "Whether to skip unresolved accounts or return an error if one or more are \
+                    missing"
+        )]
+        skip_missing: bool,
+    ) -> Result<Vec<Account>> {
+        let account_service = from_catalog_n!(ctx, dyn kamu_accounts::AccountService);
+
+        let mut res = Vec::new();
+
+        for account_id in account_ids {
+            let account_id = account_id.into();
+
+            // TODO: PERF: Vectorize resolution
+            match account_service.find_account_name_by_id(&account_id).await? {
+                Some(account_name) => {
+                    res.push(Account::new(account_id.into(), account_name.into()));
+                }
+                None if skip_missing => (),
+                None => {
+                    return Err(GqlError::gql(format!("Unresolved account: {account_id}")));
+                }
+            }
+        }
+
+        Ok(res)
+    }
+
+    /// Returns an account by its name, if found
     #[tracing::instrument(level = "info", name = Accounts_by_name, skip_all, fields(%name))]
     async fn by_name(&self, ctx: &Context<'_>, name: AccountName<'_>) -> Result<Option<Account>> {
         let account_service = from_catalog_n!(ctx, dyn kamu_accounts::AccountService);
@@ -38,6 +90,38 @@ impl Accounts {
 
         let maybe_account = account_service.account_by_name(&account_name).await?;
         Ok(maybe_account.map(Account::from_account))
+    }
+
+    /// Returns accounts by their names
+    #[tracing::instrument(level = "info", name = Accounts_by_names, skip_all, fields(?account_names, ?skip_missing))]
+    async fn by_names(
+        &self,
+        ctx: &Context<'_>,
+        account_names: Vec<AccountName<'_>>,
+        #[graphql(
+            desc = "Whether to skip unresolved accounts or return an error if one or more are \
+                    missing"
+        )]
+        skip_missing: bool,
+    ) -> Result<Vec<Account>> {
+        let account_service = from_catalog_n!(ctx, dyn kamu_accounts::AccountService);
+
+        let mut res = Vec::new();
+
+        for account_name in account_names {
+            let account_name = account_name.into();
+
+            // TODO: PERF: Vectorize resolution
+            match account_service.account_by_name(&account_name).await? {
+                Some(account) => res.push(Account::from_account(account)),
+                None if skip_missing => (),
+                None => {
+                    return Err(GqlError::gql(format!("Unresolved account: {account_name}")));
+                }
+            }
+        }
+
+        Ok(res)
     }
 }